Filter: filter id is now the first letter for most filters, unless the first character of the filter name is a digit, in which case the filter ID is the full filter name.
Filenaming: Parsing the file name for the image counter now assumes the 
image counter is UP TO FOUR digits immediately before the '.fits', 
instead of any number of digits. Still screws up for filenames ending in 
a digit (or filter names ending in a digit for PLANET reductions).

Guidezero: The calibration info for the guider stage is now stored in 
/data/guidezero, not /tmp/guidezero, avoiding the need to recalibrate 
when /tmp is cleared (on reboots)